internal firehose/util packages
