--- extension/encoding/macosunifiedloggingencodingextension/firehose_nonactivity.go
@@ -1,26 +1,0 @@
-// Copyright observIQ, Inc.
-//
-// Licensed under the Apache License, Version 2.0 (the "License");
-// you may not use this file except in compliance with the License.
-// You may obtain a copy of the License at
-//
-//     http://www.apache.org/licenses/LICENSE-2.0
-//
-// Unless required by applicable law or agreed to in writing, software
-// distributed under the License is distributed on an "AS IS" BASIS,
-// WITHOUT WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied.
-// See the License for the specific language governing permissions and
-// limitations under the License.
-
-package macosunifiedloggingencodingextension
-
-// FirehoseNonActivity represents a parsed firehose non activity entry
-type FirehoseNonActivity struct {
-	PrivateStringsSize   uint16
-	PrivateStringsOffset uint16
-}
-
-// ParseFirehoseNonActivity parses a firehose non activity entry
-func ParseFirehoseNonActivity(data []byte, flags uint16) (FirehoseNonActivity, []byte) {
-	return FirehoseNonActivity{}, data
-}
